Rewrote to better guide player and changed script names

+++ .theBeginningOfTheEnd/.floorOne/startIfStatement.sh
@@ -5,10 +5,15 @@ echo "Alright agent the gate ahead is locked. Damn, they
 	to fix the script using VIM. Do it before anymore 
 	guards show up. There are some validation scripts
 	to see if your plan was correct."
-mv ../.whileFloor/ ../whileFloor/
-mv .ifKey1.sh ifKey1.sh
-mv .ifKey2.sh ifKey2.sh
-mv .ifKey3.sh ifKey3.sh
-mv .checkIfKey1.sh checkIfKey1.sh
-mv .checkIfKey2.sh checkIfKey2.sh
-mv .checkIfKey3.sh checkIfKey3.s
+echo ""
+
+if [ -f "./.ifKey1.sh" ] 
+        then
+		if [ -f "./.checkIfKey1.sh" ]
+			then
+				mv .ifKey1.sh ifKey1.sh
+				mv .checkIfKey1.sh checkIfKey1.sh
+fi
+fi
+
+ls
